Big clean up of the code (remenants of all the re-designing).  All networking should be working except sending a file.  Remember that you need to have the server's fingerprint in your Plink/Putty cache.  During clean up I made some obvious optimizations -- example: time connecting for a dirListing is now 1/2 the time.
Also, Networking API has been modified as we discussed (I changed the names to more accurately describe the methods).  This caused a break in the cvs code which is a big no-no, but I've put in enough time in the past 3 days to deserve ya'll's forgiveness.


Process still run synchronously :-(