Plinkconnect.cpp: Changed command concatenation to use "user@host" format, since that seems to eliminate the "can't log in" error.
Networking.cpp:
Same "user@host" bit as plinkconnect
Added in a string for user home directory caching, since it's a relatively expensive operation and I'm now using it repeatedly.

Gterm.cpp:
Changed a pair of delete's in the destructor to delete[]'s.  I'm not 100% sure that this is right, but it doesn't seem to be hurting and matches up well with the allocation in the constructor.