Scrolling WORKS!  Network portion
Okay... there's LOTS of dead code in here... I'll clean it up later.  The important thing is to get it into CVS and make sure it all works together.  Color still doesn't work, but I can work on that later.

It's all about the TextManager class.  The changes basically amount to duplicating all of the text and color manipulation with calls to the TextManager object.